code/requirements.txt
adjustText==0.8
appdirs==1.4.4
appnope==0.1.3
attrs==22.1.0
autograd==1.6.2
autograd-gamma==0.5.0
backcall==0.2.0
black==24.3.0
bleach==6.1.0
certifi==2023.11.17
chardet==5.2.0
click==8.1.7
coreapi==2.3.3
coreschema==0.0.4
epydemiology==0.1.30
coverage==7.3.2
cycler==0.12.1
cryptography==46.0.3
decorator
defusedxml==0.7.1
docopt==0.6.2
docstr-coverage==2.2.0
entrypoints==0.4
et-xmlfile==1.1.0
future==0.18.3
hal-codec==1.0.2
idna
importlib-metadata==7.0.1
inflect==7.0.0
inflection==0.5.1
ipykernel==6.29.0
ipython>=7.31.1
ipython-genutils==0.2.0
itypes==1.2.0
jdcal==1.4.1
jedi==0.17.2
Jinja2==3.1.3
joblib==1.4.0
json5==0.9.14
jsonschema==4.6.0
jupyter-client==8.6.0
jupyter-core==5.5.0
jupyterlab-server==2.16.3
jupyterlab==3.5.1
kiwisolver==1.4.5
lifelines==0.27.8
llvmlite
MarkupSafe==2.1.3
matplotlib==3.6.2
matplotlib-venn==0.11.9
mistune==0.8.4
more-itertools==10.1.0
mypy-extensions==1.0.0
mysqlclient==2.1.1
nbconvert==6.5.0
nbformat==5.10.4
networkx==2.8.8
notebook==6.5.2
num2words==0.5.13
numba==0.58.1
numpy
openpyxl==3.0.10
packaging==24.0
pandas==1.5.1
pandocfilters==1.5.0
parso==0.7.1
pathspec==0.12.1
patsy==0.5.6
pexpect==4.9.0
pickleshare==0.7.5
Pillow==10.3.0
pipupgrade==1.12.0
plotly==5.18.0
pluggy==1.3.0
prometheus-client==0.18.0
prompt-toolkit==3.0.43
ptyprocess==0.7.0
py==1.11.0
pyaml==23.12.0
pycodestyle==2.11.1
Pygments==2.17.2
PyMySQL==1.1.2
pyparsing==3.1.1
pyrsistent==0.20.0
pytest==7.4.4
pytest-cov==4.1.0
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.1
quantulum3==0.9.0
regex==2023.10.3
requests==2.31.0
retrying==1.3.4
rpy2==3.5.15 
scikit-learn
scipy==1.9.3
seaborn==0.13.0
Send2Trash==1.8.2
shortuuid==1.0.11
six==1.16.0
slicer==0.0.7
SQLAlchemy==2.0.25
statsmodels==0.14.0
tableone==0.7.12
tabulate==0.9.0
terminado==0.18.0
testpath==0.6.0
threadpoolctl==3.2.0
toml==0.10.2
tornado==6.4.1
tqdm==4.66.1
traitlets==5.14.0
typed-ast==1.5.5
typing-extensions==4.7.0
UpSetPlot==0.9.0
uritemplate==4.1.1
urllib3==2.2.2
venn==0.1.3
wcwidth==0.2.6
webencodings==0.5.1
xlrd==2.0.1
zipp==3.19.2
crossref-commons